+++ lib/source/pl/lib/std/hash.cpp
@@ -20,8 +20,8 @@ namespace pl::lib::libstd::hash {
             /* crc32(pattern, init, poly) */
             runtime.addFunction(nsStdHash, "crc32", FunctionParameterCount::exactly(3), [](Evaluator *ctx, auto params) -> std::optional<Token::Literal> {
                 auto pattern = Token::literalToPattern(params[0]);
-                auto init    = Token::literalToUnsigned(params[1]);
-                auto poly    = Token::literalToUnsigned(params[2]);
+                auto init    = (u32)Token::literalToUnsigned(params[1]);
+                auto poly    = (u32)Token::literalToUnsigned(params[2]);
 
                 // Lookup table generation
                 const auto table = [&] {
